+++ native/CatapultNative/Scripts/build-app.sh
@@ -78,9 +78,9 @@ cat > "${CONTENTS_DIR}/Info.plist" <<'PLIST'
     <key>CFBundlePackageType</key>
     <string>APPL</string>
     <key>CFBundleShortVersionString</key>
-    <string>0.3.4</string>
+    <string>0.3.5</string>
     <key>CFBundleVersion</key>
-    <string>0.3.4</string>
+    <string>0.3.5</string>
     <key>CFBundleIconFile</key>
     <string>Catapult</string>
     <key>LSMinimumSystemVersion</key>